updated the getConference function to get mapfill information at the same time as the school list and to include more information in the conference info
